Refactor: AdminPlaceManager 생성
