+++ src/main/java/com/jigumulmi/admin/place/dto/response/AdminPlaceBusinessHourResponseDto.java
@@ -1,10 +1,7 @@
 package com.jigumulmi.admin.place.dto.response;
 
-import com.jigumulmi.admin.place.dto.request.AdminCreateTemporaryBusinessHourRequestDto;
 import com.jigumulmi.admin.place.dto.WeeklyBusinessHourDto;
-import com.jigumulmi.place.domain.FixedBusinessHour;
-import com.jigumulmi.place.domain.TemporaryBusinessHour;
-import com.jigumulmi.place.dto.BusinessHour;
+import com.jigumulmi.admin.place.dto.request.AdminCreateTemporaryBusinessHourRequestDto;
 import java.util.List;
 import lombok.Builder;
 import lombok.Getter;
@@ -23,23 +20,12 @@ public class AdminPlaceBusinessHourResponseDto {
     private WeeklyBusinessHourDto fixedBusinessHour;
     private List<TemporaryBusinessHourDto> temporaryBusinessHour;
 
-    public static BusinessHour fromFixedBusinessHour(FixedBusinessHour fixedBusinessHour) {
-        return BusinessHour.builder()
-            .openTime(fixedBusinessHour.getOpenTime())
-            .closeTime(fixedBusinessHour.getCloseTime())
-            .breakStart(fixedBusinessHour.getBreakStart())
-            .breakEnd(fixedBusinessHour.getBreakEnd())
-            .isDayOff(fixedBusinessHour.getIsDayOff())
+    public static AdminPlaceBusinessHourResponseDto from(WeeklyBusinessHourDto fixedBusinessHour,
+        List<TemporaryBusinessHourDto> temporaryBusinessHourList) {
+        return AdminPlaceBusinessHourResponseDto.builder()
+            .fixedBusinessHour(fixedBusinessHour)
+            .temporaryBusinessHour(temporaryBusinessHourList)
             .build();
     }
 
-    public static BusinessHour fromTemporaryBusinessHour(TemporaryBusinessHour temporaryBusinessHour) {
-        return BusinessHour.builder()
-            .openTime(temporaryBusinessHour.getOpenTime())
-            .closeTime(temporaryBusinessHour.getCloseTime())
-            .breakStart(temporaryBusinessHour.getBreakStart())
-            .breakEnd(temporaryBusinessHour.getBreakEnd())
-            .isDayOff(temporaryBusinessHour.getIsDayOff())
-            .build();
-    }
 }
