+++ src/main/java/com/jigumulmi/place/dto/BusinessHour.java
@@ -1,5 +1,7 @@
 package com.jigumulmi.place.dto;
 
+import com.jigumulmi.place.domain.FixedBusinessHour;
+import com.jigumulmi.place.domain.TemporaryBusinessHour;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.media.Schema.RequiredMode;
 import java.time.LocalTime;
@@ -25,4 +27,24 @@ public class BusinessHour {
     private LocalTime breakEnd;
     @Schema(requiredMode = RequiredMode.REQUIRED)
     private Boolean isDayOff;
+
+    public static BusinessHour fromFixedBusinessHour(FixedBusinessHour fixedBusinessHour) {
+        return builder()
+            .openTime(fixedBusinessHour.getOpenTime())
+            .closeTime(fixedBusinessHour.getCloseTime())
+            .breakStart(fixedBusinessHour.getBreakStart())
+            .breakEnd(fixedBusinessHour.getBreakEnd())
+            .isDayOff(fixedBusinessHour.getIsDayOff())
+            .build();
+    }
+
+    public static BusinessHour fromTemporaryBusinessHour(TemporaryBusinessHour temporaryBusinessHour) {
+        return builder()
+            .openTime(temporaryBusinessHour.getOpenTime())
+            .closeTime(temporaryBusinessHour.getCloseTime())
+            .breakStart(temporaryBusinessHour.getBreakStart())
+            .breakEnd(temporaryBusinessHour.getBreakEnd())
+            .isDayOff(temporaryBusinessHour.getIsDayOff())
+            .build();
+    }
 }
